refactor(linux/xdgportal): Split pipewire code into it's own file so it can be reused
Refactor pipewire_t and pipewire_display_t to it's own namespace and classes
so pipewire-based methods can share most of the logic necessary for capturing.

Also adapt portalgrab to new structure.
